refactor: social 컴포넌트 _components로 분리
