+++ src/app/social/_components/CreateSocialModalButton.tsx
@@ -1,0 +1,41 @@
+import { useAuth } from '@/providers/auth-provider/AuthProvider.client';
+import useBoolean from '@/hooks/useBoolean';
+import { useRouter } from 'next/navigation';
+import Button from '@/components/common/Button/Button';
+import { Modal } from '@/components/common/Modal/Modal';
+import CreateSocialForm from './create-social-form/CreateSocialForm';
+
+const CreateSocialModalButton = () => {
+  const { value: isOpen, setFalse: close, setTrue: open } = useBoolean();
+  const { isSignIn } = useAuth();
+  const router = useRouter();
+
+  const handleModalClick = () => {
+    if (isSignIn) {
+      open();
+    } else {
+      alert('로그인 후 이용해주세요.');
+      router.push('/auths/signin');
+    }
+  };
+
+  return (
+    <>
+      {/* 모임만들기 버튼 */}
+      {/* 버튼사이즈는 버튼 리팩토링 이후 삭제예정 */}
+      <Button size="custom" className="w-[115px]" onClick={handleModalClick}>
+        모임만들기
+      </Button>
+
+      <Modal
+        isOpen={isOpen}
+        onClose={close}
+        className="max-h-[70vh] overflow-scroll"
+      >
+        <CreateSocialForm onClose={close} />
+      </Modal>
+    </>
+  );
+};
+
+export default CreateSocialModalButton;
