+++ src/app/social/_components/detail/[socialId]/NotFoundRedirect.tsx
