+++ src/app/social/_components/detail/[socialId]/type.ts
