+++ src/app/social/_components/social-list/SocialInfiniteScrollFeed.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import { SocialListProps } from '../type';
+import { useGetSocialList } from '@/hooks/api/social';
+import Observer from '@/components/common/Observer/Observer';
+import SocialListGrid from './SocialListGrid';
+
+const SocialInfiniteScrollFeed = ({ filter }: SocialListProps) => {
+  const { data, isLoading, fetchNextPage, isFetchingNextPage, hasNextPage } =
+    useGetSocialList(filter);
+
+  const socialList = data?.pages.flat() ?? [];
+
+  return (
+    <>
+      <SocialListGrid socialList={socialList} isLoading={isLoading} />
+      <Observer
+        enabled={hasNextPage && !!socialList.length}
+        onIntersect={() => {
+          if (hasNextPage && !isFetchingNextPage) {
+            fetchNextPage();
+          }
+        }}
+      />
+    </>
+  );
+};
+
+export default SocialInfiniteScrollFeed;
