+++ src/app/social/_components/type.ts
